+++ pkg/compiler/lib/src/js_backend/native_data.dart
@@ -38,9 +38,6 @@ class NativeBasicDataBuilder {
   /// The JavaScript members implemented via typed JavaScript interop.
   final Map<MemberEntity, String> _jsInteropMembers = {};
 
-  /// The JavaScript interop members annotated with `@ObjectLiteral`.
-  final Set<MemberEntity> _jsInteropObjectLiterals = {};
-
   /// Sets the native tag info for [cls].
   ///
   /// The tag info string contains comma-separated 'words' which are either
@@ -110,8 +107,7 @@ class NativeBasicDataBuilder {
 
   /// Marks [element] as an explicit part of js interop and sets the explicit js
   /// interop [name] for the member [element].
-  void markAsJsInteropMember(MemberEntity element, String name,
-      {required bool isJsInteropObjectLiteral}) {
+  void markAsJsInteropMember(MemberEntity element, String name) {
     assert(
         !_closed,
         failedAt(
@@ -119,7 +115,6 @@ class NativeBasicDataBuilder {
             "NativeBasicDataBuilder is closed. "
             "Trying to mark $element as a js-interop member."));
     _jsInteropMembers[element] = name;
-    if (isJsInteropObjectLiteral) _jsInteropObjectLiterals.add(element);
   }
 
   /// Creates the [NativeBasicData] object for the data collected in this
@@ -134,8 +129,7 @@ class NativeBasicDataBuilder {
         _jsInteropClasses,
         _anonymousJsInteropClasses,
         _staticInteropClasses,
-        _jsInteropMembers,
-        _jsInteropObjectLiterals);
+        _jsInteropMembers);
   }
 
   void reopenForTesting() {
@@ -174,9 +168,6 @@ class NativeBasicData {
   /// The JavaScript members implemented via typed JavaScript interop.
   final Map<MemberEntity, String?> _jsInteropMembers;
 
-  /// JavaScript interop constructors annotated with `@ObjectLiteral`.
-  final Set<MemberEntity> _jsInteropObjectLiterals;
-
   NativeBasicData(
       this._env,
       this._isAllowInteropUsed,
@@ -185,8 +176,7 @@ class NativeBasicData {
       this._jsInteropClasses,
       this._anonymousJsInteropClasses,
       this._staticInteropClasses,
-      this._jsInteropMembers,
-      this._jsInteropObjectLiterals);
+      this._jsInteropMembers);
 
   factory NativeBasicData.fromIr(
       KernelToElementMap map, IrAnnotationData data) {
@@ -197,7 +187,6 @@ class NativeBasicData {
     Set<ClassEntity> anonymousJsInteropClasses = {};
     Set<ClassEntity> staticInteropClasses = {};
     Map<MemberEntity, String?> jsInteropMembers = {};
-    Set<MemberEntity> jsInteropObjectLiterals = {};
 
     data.forEachNativeClass((ir.Class node, String text) {
       nativeClassTagInfo[map.getClass(node)] = NativeClassTag(text);
@@ -217,17 +206,13 @@ class NativeBasicData {
         staticInteropClasses.add(cls);
       }
     });
-    data.forEachJsInteropMember((ir.Member node, String? name,
-        {required bool isJsInteropObjectLiteral}) {
+    data.forEachJsInteropMember((ir.Member node, String? name) {
       // TODO(49428): Are there other members that we should ignore here?
       //  There are non-external and unannotated members because the source code
       //  doesn't contain them. (e.g. default constructor) Does it make sense to
       //  consider these valid JS members?
       if (memberIsIgnorable(node)) return;
       jsInteropMembers[map.getMember(node)] = name;
-      if (isJsInteropObjectLiteral) {
-        jsInteropObjectLiterals.add(map.getMember(node));
-      }
     });
 
     return NativeBasicData(
@@ -238,8 +223,7 @@ class NativeBasicData {
         jsInteropClasses,
         anonymousJsInteropClasses,
         staticInteropClasses,
-        jsInteropMembers,
-        jsInteropObjectLiterals);
+        jsInteropMembers);
   }
 
   /// Deserializes a [NativeBasicData] object from [source].
@@ -261,7 +245,6 @@ class NativeBasicData {
     Set<ClassEntity> staticInteropClasses = source.readClasses().toSet();
     Map<MemberEntity, String?> jsInteropMembers = source
         .readMemberMap((MemberEntity member) => source.readStringOrNull());
-    Set<MemberEntity> jsInteropObjectLiterals = source.readMembers().toSet();
     source.end(tag);
     return NativeBasicData(
         elementEnvironment,
@@ -271,8 +254,7 @@ class NativeBasicData {
         jsInteropClasses,
         anonymousJsInteropClasses,
         staticInteropClasses,
-        jsInteropMembers,
-        jsInteropObjectLiterals);
+        jsInteropMembers);
   }
 
   /// Serializes this [NativeBasicData] to [sink].
@@ -289,7 +271,6 @@ class NativeBasicData {
     sink.writeClasses(_staticInteropClasses);
     sink.writeMemberMap(_jsInteropMembers,
         (MemberEntity member, String? name) => sink.writeStringOrNull(name));
-    sink.writeMembers(_jsInteropObjectLiterals);
     sink.end(tag);
   }
 
@@ -396,8 +377,6 @@ class NativeBasicData {
         map.toBackendClassSet(_staticInteropClasses);
     Map<MemberEntity, String?> jsInteropMembers =
         map.toBackendMemberMap(_jsInteropMembers, identity);
-    Set<MemberEntity> jsInteropObjectLiterals =
-        map.toBackendMemberSet(_jsInteropObjectLiterals);
     return NativeBasicData(
         environment,
         isAllowInteropUsed,
@@ -406,8 +385,7 @@ class NativeBasicData {
         jsInteropClasses,
         anonymousJsInteropClasses,
         staticInteropClasses,
-        jsInteropMembers,
-        jsInteropObjectLiterals);
+        jsInteropMembers);
   }
 }
 
@@ -625,10 +603,6 @@ class NativeData implements NativeBasicData {
   Map<MemberEntity, String?> get _jsInteropMembers =>
       _nativeBasicData._jsInteropMembers;
 
-  @override
-  Set<MemberEntity> get _jsInteropObjectLiterals =>
-      _nativeBasicData._jsInteropObjectLiterals;
-
   /// Returns `true` if [element] has an `@Anonymous` annotation.
   bool isAnonymousJsInteropClass(ClassEntity element) {
     return _anonymousJsInteropClasses.contains(element);
@@ -637,11 +611,6 @@ class NativeData implements NativeBasicData {
   /// Returns `true` if [element] has an `@staticInterop` annotation.
   bool isStaticInteropClass(ClassEntity element) {
     return _staticInteropClasses.contains(element);
-  }
-
-  /// Returns `true` if [element] has an `@ObjectLiteral` annotation.
-  bool isJsInteropObjectLiteral(MemberEntity element) {
-    return _jsInteropObjectLiterals.contains(element);
   }
 
   @override