+++ pkg/compiler/lib/src/kernel/native_basic_data.dart
@@ -95,9 +95,7 @@ class KernelAnnotationProcessor {
             /*reporter.reportErrorMessage(
                 function, MessageKind.JS_INTEROP_NON_EXTERNAL_MEMBER);*/
           } else {
-            _nativeBasicDataBuilder.markAsJsInteropMember(function, memberName,
-                isJsInteropObjectLiteral:
-                    annotationData.isJsInteropObjectLiteral(memberNode));
+            _nativeBasicDataBuilder.markAsJsInteropMember(function, memberName);
             // TODO(johnniwinther): It is unclear whether library can be
             // implicitly js-interop. For now we allow it.
             isJsLibrary = true;
@@ -142,8 +140,7 @@ class KernelAnnotationProcessor {
               // TODO(johnniwinther): The documentation states that explicit
               // member name annotations are not allowed on instance members.
               _nativeBasicDataBuilder.markAsJsInteropMember(
-                  function, memberName,
-                  isJsInteropObjectLiteral: false);
+                  function, memberName);
             }
           }
         });
@@ -160,8 +157,7 @@ class KernelAnnotationProcessor {
             // TODO(johnniwinther): The documentation states that explicit
             // member name annotations are not allowed on instance members.
             _nativeBasicDataBuilder.markAsJsInteropMember(
-                constructor, memberName,
-                isJsInteropObjectLiteral: false);
+                constructor, memberName);
           }
         });
       }